init-db/requirements.txt
# init-db/requirements.txt

weaviate-client
requests
python-dotenv